+++ CHANGELOG.md
@@ -9,12 +9,22 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Features
 
+### Fixes
+
+### Breaking
+
+## [2.104.0] - 2024-12-23
+
+### Features
+
 - program: pyth lazer integration ([#1361](https://github.com/drift-labs/protocol-v2/pull/1361))
 - program: add ix to log user balances ([#1366](https://github.com/drift-labs/protocol-v2/pull/1366))
 
 ### Fixes
 
-program: fix force delete user for token 2022 ([#1358](https://github.com/drift-labs/protocol-v2/pull/1358))
+- program: fix force delete user for token 2022 ([#1358](https://github.com/drift-labs/protocol-v2/pull/1358))
+- program: fix liquidating dust prediction mkt position ([#1397](https://github.com/drift-labs/protocol-v2/pull/1397))
+- program: spot market decimals under 6 precision fixes ([#1399](https://github.com/drift-labs/protocol-v2/pull/1399))
 
 ### Breaking